refactor drawing manager package name
