+++ artmaker/src/main/java/io/artmaker/drawing/DrawingManager.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.artmaker.export
+package io.artmaker.drawing
 
 import android.graphics.PointF
 import androidx.compose.runtime.mutableStateListOf
